Refactor to package structure
