+++ zenday/mappers/__init__.py
@@ -1,0 +1,1 @@
+from .kroger import map_kroger_to_zenday
